src/item.py
from csv import DictReader


class Item:
    """
    Класс для представления товара в магазине.
    """
    pay_rate = 1.0
    all = []

    def __init__(self, name: str, price: float, quantity: int) -> None:
        """
        Создание экземпляра класса item.

        :param name: Название товара.
        :param price: Цена за единицу товара.
        :param quantity: Количество товара в магазине.
        """
        self.__name = name
        self.price = price
        self.quantity = quantity
        Item.all.append(self)

    def calculate_total_price(self) -> float:
        """
        Рассчитывает общую стоимость конкретного товара в магазине.

        :return: Общая стоимость товара.
        """
        return self.price * self.quantity

    def apply_discount(self):
        """
        Применяет установленную скидку для конкретного товара.
        """
        self.price = self.price * self.pay_rate
        return self.price

    @property
    def name(self):
        return self.__name

    @name.setter
    def name(self, name):
        if len(name) <= 10:
            self.__name = name
        else:
            self.__name = name[:10]

    @classmethod
    def instantiate_from_csv(cls, file):
        with open(file, encoding='Windows-1251') as file:
            reader = DictReader(file, delimiter=',')
            for row in reader:
                Item(row['name'], float(row['price']), int(row['quantity']))

    @staticmethod
    def string_to_number(name):
        if type(name) is str and name.isdigit:
            return int(float(name))

    def __str__(self):
        return f"{self.__name}"

    def __repr__(self):
        return f"Item{self.__name, self.price, self.quantity}"






